CE-6909: review: fix concurrency issue that may lead to multiple shaders having the same name
The logic for making our shader names unique relies on the compute method of the arnold material
node seeing all the other existing nodes in the scene. However, since the compute method of several
nodes can be executed in parallel, two or more arnold material nodes may compute the same 'unique'
name for their shaders, thereby breaking uniqueness. To fix this issue, we explicitely set the
scheduling type of the arnold material node as globally sequential node.

Moreover, the assembled MEL script is executed asynchronously. We specifically execute the creation
of the shader node synchronously such that the compute method of following arnold material node
updates is guaranteed to see these nodes and consider them for finding unique node names. However,
this is fragile and may be a source of errors, since the method used is not thread safe.