Update UI after attribute changes
This updates the attribute slider to the currently used value in the corresponding prt attribute.
I.e. update the attribute when it's was reset to the default value.
It was also possible that an attribute did not properly update after UI changes (e.g. when changing a float attribute via the field instead of the slider) this should fix that issue.