Check entire history for serlio node
When an initial shape is generated without constuction history and a serlio-node is appended the geometry node gets duplicated as a placeholder input.
Therefore its possible that a geometrynode returned by "collectInitialShapes" is actually one of these duplicate nodes. In order to avoid generating additional serlio nodes for these duplicates, we have to check the future and past connections.