Remove declaration for the Shading engine
This caused a warning in the mel console.
The shading engine variable is already used (and thereby declared) earlier in synchronouslyCreateShadingEngine().
Since the synchronous code is executed before the other lines, removing these optional declarations fixes these warnings.
Another alternative solution would be to force the variable to be declared synchronously with MELScriptBuilder.executeSync(string output)