Only allow installation of newer serlio versions.
If the serlio version matches we will now show repair/delete dialog instead of just "overriding" old installations of the same version.
If we try to install an older version of serlio on a newer one, we will now get an error asking us to uninstall the newer version first.